graph: Fix visualise_hierarchy's edges
Before, they were hidden underneath the nodes. Now, they're above and
pretty obvious.

Unfortunately, I've tried a million things to make the arrows go to the
edges of the nodes instead of the centres like it is right now, but
nothing worked.

This is pretty much a debug output, so I'll leave it like this.