gpt(extract_graph): Refactor relation between extract and graph
The idea as that extract_graph should use nx.DiGraph as little as
possible, at least directly, so it can use strict typing, and leave
networkx confined to a single file that can use basic typing.

It helps with naming stuff when I limit the graph import to a named
import instead of importing each thing separately. Unfortunately, I
can't think for a better name than "graph" for the module (no, I'm not
calling it "graphutils"), so I have to rename it to something that won't
clash.