gpt(extract_graph): Add paper approval classification based on the graph
This commit does way too much because I wasn't disciplined while doing
it, but we have:

- Add method to Paper to evaluate whether it would be approved. There
are two strategies for now, with an explicit default.
- Add classification function and prompts
- Add CLI stuff for the classification (it has to be enabled by a flag,
and there are separate prompts and keys for it)
- Refactor the CLI stuff to have the main function in a `run`
subcommand, and use the `prompts` subcommand to list currently available
prompts for both extract and classify
- Refactor GptResult dataset to accept a T, so I don't have to duplicate
the results for extract and classify. The GPT run functions are
duplicated because I expect the classification one to end up different,
even if it's just a clone for now
- Save the classification results (input data plus y_true and y_pred) as
JSON, together with the classification metrics