gpt(extract_graph): Add graph verification
The other one was proper graph math from networkx, which is still in
use, since it makes it easy to identify cycles.

This one is directly checking the invariants we want to hold. I could've
implemented these in networkx too, but I felt it was better to have this
here.

I'm still not happy with the "return error string" idea, but since it's
only for debugging (currently), it should be enough.