tasks(setup): Add uv to path after installing it
uv's installation automatically adds ~/.cargo/bin to the path through
the shell's configuration file, but that doesn't do anything for the
current session. This means that people who are installing uv through
the `setup` task have the rest of the uv stuff fail.

Instead, the script now manually sources the environment file. This
seems to work.