s2orc(ss): Read paper information from JSON and save full output
Instead of a text file with one title per line, now we have a JSON with
the first author too. It's not used here, but the Crossref script can
use both.

Also, we were only storing the valid results, but I also want to save
the full output in case that becomes useful.