archive: Add script to analyse approval decision vs ratings
Currently only checks the strategies to convert the numeric ratings to a
binary label vs the provided binary approval decision.

Conclusion: MEAN is a better strategy than MAJORITY (better F1 and
accuracy). Both of them have ~86% F1, so I'm confident that the ratings
and the approval decision are aligned. That was already supposed to
happen, but it's good to have some confirmation.