archive: Change data structure from S2 title fuzz script
Instead of creating another type of file with just title/query/fuzz,
instead copy all the input data and just add the fuzz_ratio to it.

In addition to that, accept a flag to filter the papers by a minimum
fuzzy ratio, and create a separate file for the entries above that
threshold.