asap(reference): Add the original S2 result title to references output
Before, we only had the title on the ASAP side of the references.
However, these can differ from the S2 title (hence the fuzzy matching).
It can still be useful to have the original title in case we need to
match the same S2 data to multiple papers in ASAP.

For now, I'm keeping the S2 data nested in the ASAP output because I
think it makes the most sense, but this allows identifying different
papers referring to the same S2 paper.