asap(extract): Expand context sentences within their paragraphs
We need more information to classify the contexts since just the
sentences provided by ASAP are not enough in many cases. So we try to
get more sentences surrounding the contexts. However, we cannot include
sentences that contain other citations.

The current method of doing this is facing issues when splitting the
paragraph into sentences, so the test currently fails.