deps: Add en-core-web-sm as explicit dependency
This also means removing `pip` as a dependency, since it's only used to
install the module.

The old version was working fine. However, I wanted to do this because
`uv sync` would remove packages that aren't in pyproject.toml, so it
would uninstall the spacy/pip-installed version every time. With this,
it's explicit.