gpt(run_gpt): Add types that bundle the output with its prompt
I'm still not 100% sure of the interaction between GPTResult, Prompt,
PromptResult, etc. Right it's a bit nested, with GPTResutl being result
+ cost, and PromptResult being result + prompt.

This works for now because PromptResult is per item, and GPTResult works
both for a single item and for the collection. Still, I'm not happy and
might refactor this in the future.

Also, GPTResult is a dataclass because it's just a bundle and doesn't
need validation/serialisation. On the other hand, I need Prompt to be a
BaseModel because I'll persist it later. I don't think I'm persisting
PromptResult, but it feels a little weird not to have it be a BaseModel
too.