gpt(classify_contexts): Adopt PromptResult
I'm not happy about this one because for each paper, there will be many
model requests for different contexts, but the classification function
is setup so that we return the full paper (with its classified contexts
inside) as the output.

For now, I'm just greating one big prompt with the concatenation of all
the context prompts, but I don't like it.