gpt: Fix logging
I'm using hardcoded names for the loggers instead of just __name__
because the full name is too long. However, I need to explicitly make
the classify_contexts.py and extract_graph.py loggers subloggers of the
cli.py one to make this work.

Also, we need to explicitly initialise the 'gpt' logger in extract_graph
so the logger in evaluate_graph is used too. If I use the extract_graph
logger directly, evaluate_graph's isn't initialised.

This sounds a bit convoluted, but it's not that bad. I changed the
`setup_logging` function to accept the logger name, so I just use
`setup_logging('gpt')` everywhere and it works.