gpt(classify_contexts): Add classification evaluation
Now we convert the input data's trinary polarity (including "neutral")
to binary ("neutral" becomes positive), save both pred and gold in the
output and calculate binary evaluation metrics.

There's one problem: I'm asserting the data entries include the relevant
annotated polarities in the middle of the loop, but I should think of
something to do when the annotation is unavailable. I feel like I
shouldn't outright reject the data. Maybe let the 'gold' be binary and
handle that later?