asap(model): Add property to PaperReference for always-valid contexts
Because of how I set up the types, the ASAP pipeline doesn't create
JSONs with the `contexts_annotated` field. However, for it to work with
the `classify_contexts` script, we need something to iterate through,
even if the actual polarity is absent. This property creates that at
runtime if required.