gpt(classify_contexts): Default to "sentence" prompt
The default setting is now:
- prompt: 'sentence' (only the context itself)
- context: 'original' (not expanded)
- model: gpt-4o-mini (gpt-4o is only marginally better, <1% F1, but 20x
the price and 3x the time

I.e. the simplest thing possible is pretty good. Hanqi thought this from
the beginning, but I was sceptical.

The key point here is that when annotating, I classify into
positive/neutral/negative. 'neutral' cases are the ones where I don't
think there's an obvious positive aspect. However, Hanqi thinks that
anything not explicitly negative should be positive, so in practice, I
convert 'neutral' to 'positive' (see asap.model.ContextPolarityBinary).

In the end, this means the annotated data is >85% positive, and the
classifier achives >95% F1, which is pretty good.

This is good enough for now. I might come back and tackle context type
again in the future.