deps: Clean-up unused and indirect dependencies
Removed:
- spacy: I'm not using it to split paragraphs since we're not doing
context expansion anymore, so it goes
- en-core-web-sm: this is the spacy model I was adding explicitly, so it
goes too
- ratelimit: I looked and it seems like this was added pre-emptively and
never used. Probably because I used habanero for Crossref and
semanticscholar for Semantic Scholar, which should take care of this, or
when I did query the API direclty with aiohttp, I used a semaphore to
avoid multiple simultaneous tasks, along with directly watching out for
429 codes.

Added:
- click: indirect dependency from typer that I'm importing directly so I
can use click.Choice when prompting of an annotation. There doesn't seem
to be a proper typer way of doing this, but click.Choice and
typer.prompt work well.