gpt(classify_contexts): Enable reading results from previous run
Since we're saving interim results after each paper completes, we can
now get a path to such interim results for a new run and avoid querying
the same data.

This is all because there's a bug that's killing the process in the
middle of classifying, and since I can't figure out why, I'll do this
incrementally.