gpt(run_gpt): Add rate limiter to run_gpt_async
OpenAI's chat parameters are a nested mess of different TypedDicts, and
they don't make it easy to use them in practice. For example, the types
are split across several files.

For now, I'm just using Any here so it type checks. I already have to
add `type: ignore` for openlimit anyway since it doesn't use a single
type hint.

If this works well enough, I might copy the code and add type hints.
`openlimit` is GPL v3, which is great.