gpt(classify_contexts): Add warning to use about connection errors
This whole thing seems to work, most of the time. I'm getting a good
number of error messages from the OpenAI client. I'm not sure what they
mean, but they're likely from the high volume.

The solution here is to just let it rip, let it finish, and retry until
we have all items.