gpt(extract_graph): Add `GPTGraphStrict` output type
Aims to improve model output by making explicit entity types and
relationship restrictions. For example. it is no longer possible to
create relationships between title and claims, a common problem with
GPT-4o-mini.

The risk is the model generating malformed JSON output that cannot be
parsed into this type, but we'll see.