util: Add TopK collection
As I add more items to it, it only keeps the top K of them by value. The
items are compared using the type's `__lt__` method.

I tried using a comparison function instead. However, heapq uses tuples
to decide the order, and if there's a tie in one item, it moves on to
the next. This means if there is ever a tie in between items, it could
go on until the original item, and if that's not comparable, it throws
an Exception.

Instead, I'm requiring the input type to be comparable, which is a
little problematic, but easier for the class.