git: Fix get_remaining items
- Function now returns both done and remaining items
- Remaining items are processed then combined with the done items
afterwards
- Add title and abstract to Graph so I can use to properly get the id of
the underlying paper. I was relying on the generated title, but doesn't
work properly.
- Take `key` functions for `continue` and `original` data. I was trying
to use a protocol for an `id` property, but that doesn't match well with
pydantic's BaseModel since you can't combine a Protocol and an ABC. I
thought about creating my own base model with an `id` property, but
quickly gave up on the idea.