s2orc: Use urllib.parse.urlparse directly instead of wrapper
I wrote the wrapper because I thought there was some type-checking issue
going on because pyright couldn't infert the type of the `path` member
of the result.

The issue was this: `urlparse` takes either a URL as str or bytes, and
the output type depends on that. However, the `url` I was passing in
came from a JSON and was `Any`. This means that pyright couldn't figure
out why overload to use.

The solutions were either:
- Assert the type was string (runtime, dangerous)
- Use `typing.cast` to tell the type system I knew it was a string
(verbose)
- Just `str(url)` (runtime, not particularly dangerous, concise)

I took the last one.