Use `__name__` instead of explicit package name for logging
I'm using the full name for everyone to make sure that hierarchical
logging setup works, so I might as well use `__name__`.

And to make it better, I'm using the root package from `__name__` in the
`setup_logging` function instead of manually `paper`, in case the name
changes one day and I forget to change this here.