gpt: Fix logger initialisation
When running the scripts directly, `__name__` is `'__main__'`.
Initialising a logger with `__name__` will not initialise the `paper`
package root logger, so not everything is properly logged.

I want the scripts to work if executed standalone, but that means
initialising both the `__main__` logger (for the script's own logs) and
the `paper` (package root) logger (for the imported packages logs).
`setup_logging` is now explicitly initialising both, with no
"auto-discovery" going on. Non-main packages will be children of
`paper`, and the `__main__` package has its own initialisation.

This is a little awkward, but I couldn't find better ways to do it.

The cleanest fix would be not to support running standalone scripts at
all, but I'm not sure whether I want to do that.