external(s2_references): Script used to get reference information
This is the one originally used to get abstract information on the
papers cited by those in ASAP. The reason I'm copying it is because it's
not very reliable. It doesn't really care about rate limits or the 429
error in particular; it just keeps trying. This means that some papers
will just fail, which is fine when you're querying 15k references, but
when I want precise information about the main papers, it isn't enough.