external(s2_info,s2_model): Split info output models and move outside
I split the outputs of `info` into two: one that maybe has an S2 (the
output of the queries), and one that definitely does. This is useful
because most places will use the latter, and I don't have to keep doing
`if paper.s2` everywhere.

They're moved to the semantic_scholar_model modules so that the
`_recommended` script can import from there (I don't like to import from
scripts rather than modules).