gpt(cli): Improve help text for subcommands
For subcommands that have a `run` subsubcommand, the help text is like
this:
- In the `gpt` command, the subcommand gets a short message saying what
it does
- In its subcommand (e.g. `gpt terms`), the description is the summary
of the module docstrings, the `run` commnad gets a similarly short
message, and `prompts`'s message is fixed
- In the subsubcommand (e.g. `gpt terms run`), the description is the
full module docstring.

Most commands follow this exact format, so I even created a collection
and loop for them. `tokens` is different since it's not a command with
`run/prompts`, so it's separate.

PS: This would all be so easier if I were using typer from the
beginning, but I cba to refactor all of that now. Although this was
already quite a long refactor...