Use Record type to fix `get_remaining_items`
The input types (raw data and intermediate results) need both to have an
`id` property and also be subclasses of `BaseModel`. The former for the
deduplication logic to work, and the latter for `PromptResult` to work,
as it needs to build a schema and that only works with concrete types
(no Protocols).

The input types of `get_remaining_items` (raw data and intermediate
results) need to fulfil two requirements:

- Have an `id` property that returns a `Hashable`: used to identify an
instance during deduplication when determining which instances need to
be processed.
- Be concrete types: they're embedded in `PromptResult`, which is a
`BaseModel`, so it needs to be able to generate a schema for all its
fields. If the type is a Protocol - even if @runtime_checkable - it
can't generate a schema, regardless of any options.

This means T (and U, by extension) has can't be a Protocol, so we need
an abstract base class to define the `id` property. Instead of doing
that to every type, I decided to create my own base class for future
models, `Record`, which has the `id` property, is a `BaseModel` subclass
and has the commonly useful settings for `ConfigDict`: `frozen=True,
populate_by_name=True`. I only set this as the base class for the cases
needed for the program to type-check and run, but I'll be using it for
future new types.

PS: I put this in `util` because I couldn't think of a better place, and
I thought it was too small to warrant a file just for it. I might be
wrong on that. Also, `util` has become what every `util` model is
destined to be: a bundle of unrelated stuff in a single place. I'll
eventually refactor this, but I'm not quite sure yet. One solution is to
create an `util` package with related things in different files,
respecting cases where a short file might have to do.