gpt(annotate_terms): Allow multiple output types
This is a little different from the others, which have the structured
outputs as intermediate types, which are all transformed to a single
type right after the GPT call.

Here, we don't have such unified type yet since I haven't finalised the
term graph representation. I don't even know if I want to output the
graph from here - probably not. When I finally stabilise the graph, it
will depend on a certain kind of output, and that will be created here.
Until then, I'll allow different kind of outputs shapes.

This means that I need a concrete `type_` object that I pass through
everything so the functions are all generic on that. This only needed
because I need the `type_` parameter to bind the `T` type variable. I
only really use the `type_` when actually running the GPT call. This
also means an `empty` method since I can't construct the types without
knowing what they are.