gpt(annotate_terms): Use a single --log flag for stats
If it's not given, no log is shown.
If it is given, but it's empty or `simple`, only shows counts.
If it is given as `detail`, show the full details.
Any other value is an error.

This is suprisingly easy to get working on argparse, and apparently
impossible to do on typer.