gpt(model,annotate_terms): Define `S2Paper` with required fields
The `external_data` version of this has a bunch of nullable fields in
case the API returns missing information, but as input for annotation,
we need the full thing.

The only nullable field now is `tldr`. `tldr` is expected to be null in
a sizable amount of papers, and we're not using it for anything.