gpt(annotate_terms): Split abstract into "context" and "target"
This is done by a separate prompt in the core of the module, in the
`_annotate_paper_single` function. We make two requests: one for term
extraction and relation classification, another to split the abstract,
then collect everything as a `PaperAnnotated`.

These will be used for the Semantic Neighbour graph, as they're part of
the prompt for the embeddings.