external(s2_info): Remove old info script and rename references
`references.py` (before renaming) has both `main` and `references` mode,
so it can do the job of the previous `info.py` through the `main` mode.
As it was redundant, I'm removing it and renaming `references.py` to
something more generic.

Also updates the docstring on the new `info.py` to better explain both
modes.