scimon(kg): Build KG graph
Ignores the nodes in GPTTerms, using only the relations with the heads
as source nodes, and the tails as target nodes. We pre-process the heads
before adding to the graph. Graph is represented as an edge list. We
keep the list of "nodes" together with pre-calculated embeddings.

When querying the graph, we take the text and pre-process it. If there's
an exact match in the graph, we return its neighbours. If not, we get
the most similar neighbour by embedding and return the neighbours for
that. We keep track of the source in the query result.