docs: Add `clidoc` command to generate gpt CLI documentation
It's quite neat: it parses the help text and generates a full markdwon
file with the documentation. I'm not sure how useful it is (I think I'd
rather keep running the commands with `--help`), but it's low effort, so
why not.

The only concern is that this might fall out of sync if I forget to run
the generation command, and I'm not sure it's worth complicating the
workflow to automatically update the file. I guess I'll just try to
remember.