paper: Make all paper IDs strings
I was using built-in `hash` to hash title+abstract, but it turns out it
differs from run to run, which is not great. I'm using a proper hash
now.

I'm also making the `Record.id` property return a string instead of
`Hashable` just in case I forget this and use an int again.