gpt(annotate_paper): Allow both ASAP and S2 papers as input
Uses a type alias, PaperToAnnotate, to asap.Paper and gpt.S2Paper. All
member accesses must be through common items between them. The user
needs to give the type in the CLI so the script knows what to use to
load the data.

I'm not defining a new type with the common members because I need the
full data. I'm going to run this annotation script on the ASAP data, and
I need all the data from ASAP there.