gpt(eval_full): Fix default prompt keys
Since I'm not using enums, there's nothing to tell me that they drifted
from the real names, so I have to manually fix this.

Also, the input/output paths are now required options instead of
positional arguments. I read about this on clig.dev - prefer flags to
args. It makes sense when you have as many parameters as these scripts
tend to have, so I might subscribe to that a bit more.