s2: Flatten the external_data/semantic_scholar namespace
"Flat is better than nested". Namespaces are not hierarchies. It was
like this because we had both Crossref and Semantic Scholar, but we
don't anymore.

The "cloud" thing makes sense in the same packge as the S2 API scripts
because they're basically the same thing (fetch information from S2)
with different targets (Athena/S3 vs local files).