scimon(embedding): Add `batch_encode` to Encoder class
The main benefit is that I can run a tqdm progress bar on top of the
chunks so I have some sense of progress. I tested some batch sizes, and
128 seems to be a good number. Increase this didn't make any difference
in runtime.

This progress bar is optional, in case we need to process this
differently in the future. However, the current use case has the
progress bar on for everything, except Citations, since the embeddings
are already "batched" in small groups there (the unit of work is the
paper).