scimon(embedding): Remove `multi` method
It's meant to be used for multiple GPUs, which we don't have. The
performance of the standard `encode` method is basically the same if we
use batches, so I'm removing this.

It has the benefit of not needing to init/close the pool, which means
the Encoder doesn't have to be a context manager anymore. This
simplifies code all over.