gpt(eval_scimon): Query graphs in process executor; group results
I was running the graph query directly, but it's CPU-bound and was
blocking the event loop. Now, I'm running it in a separate process,
which means I have to load the graph there, not in the main function.

This is because I can't send the Graph across process boundaries, and
Python's shared memory stuff is way too complex.

I also tried running the graph query in parallel, but that means loading
the graph once per process, which used too much memory and I gave up.

Since I'm doing the querying in one go, I also needed to wrap the paper
and the query results in a single object, which I now send around
instead of just the paper.