gpt(eval_scimon): Use pre-graph-queried data as input
Instead of running the queries every time, pre-compute the terms so the
only thing the eval script does is take the data and send to GPT. This
saves some time and reduces the complexity of the script.

This meant I had to change the dependency file/graph, update the
processing script, and move some classe around to avoid circular
dependencies.