Add 'authors' field to ASAP papers
This means changing the various representations of ASAP papers across
the codebase[1] and fixing the places where the objects are constructed.
This all went remarkably well, all thanks to Pydantic and static typing.

[1] There are way too many. Some are derivations (e.g. ASAP paper +
S2-augmented references), but there is asap.Paper and gpt.S2Paper, which
probably don't need to exist. On the other hand, mabye it's fine if two
modules have their own representation for what they need/want? Although
I'd expect the `asap` and `s2` definitions to be global. Anyway, this
wasn't so bad, so I'm kicking this can further down the road.