Change S2 `title_query` to `title_asap`
Every time I see `title_query` I have to remember that it's "the title
from the ASAP data that I used to query the S2 API, but it's not the
title in the S2 data". `title_asap` makes that clear.

This is an alias of the data rather than a new field because I don't
want to change the S2 data scripts since re-running them would take so
long.