deps: Replace Pandas with Polars
All my uses of Pandas were to get descriptive stats from a sequence of
numbers, and Polars slots right in. If I need more in-depth stuff in
the future, I'll use Polars as much as I can.

The main reason for this is that Polars has proper type hints and
Pandas doesn't. I haven't used Polars in earnest enough to make any
comments about the API, but everyone seems to hate Pandas's API and at
least tolerate Polars's.