peter(citations): Decouple Polarity enum from other packages
I was borrowing the `asap.ContextPolarity` enum, but it was the last
non-CLI connection to the other packages, so I removed it. It's a little
cumbersome having to define both the Protocol (which is itself weird)
and the concrete Enum, but it seems to work, and it it's not _that_ much
boilerplate.

Static typing haters are rolling in their graves after this, of course.