tasks: Remove `setup` tasks and update README
The task was finicky and was there for less technical people. Now that I
don't really care about that anymore, it's better to just have the
instructions in the README and be done with.

Technically, you don't even need to run `uv sync`. What's great about
`uv` is that, if you have it installed, you can just go into a new repo
and run `uv run ...`, which will set up everything for you. The main
concern there is having people use pre-commit, but if they don't, CI
will catch mistakes anyway.