gpt(model): Decouple `PaperRelatedSummarised` from `peter`
The `peter` module depends on types from `gpt.model`, so I was getting a
circular import. I fixed it by removing the data inheritance (probably a
good idea anyway for a type as small as this) and importing only `peter`
for type-checking.

I don't like to use `if TYPE_CHECKING` because it looks like a horrible
hack, but it was required here. Either this or somehow refactoring this
into a separate module, but that's totally unnecessary, so hack it is.