gpt(eval_peter): Add prompt without related papers for testing
Just running the same script with this prompt performs better (66%
accuracy with 10 demonstrations and GPT-4o) than the other prompt with
the related papers.

This is mostly to show that the problem isn't the code, it's the prompt.
I'll follow this up with changes to the summarisation and the 'simple'
prompt to try to improve this.