s2: Rework `info`/`recommended` types
They fell out of sync, and the output of `info` was no longer compatible
with the expected input of `recommended`. It turns out that I couldn't
share the code between main and references in `info` like I was doing,
since different tools expected different formats for each, so I split it
into two commands.