tests: Move ASAP pipeline test to PeerRead
The existing GPT and PETER scripts have all been changed to support
PeerRead, so this is just moving the existing ASAP pipeline to PeerRead.
No change in functionality.

Except one: this doesn't always automatically download the PeerRead
dataset, only if it cannot be found. The ASAP version always downloaded
it from scratch. This is because the dataset is on GitHub and is quite a
bit larger (~300MB ASAP to ~2GB PeerRead), so I don't want to do it
every time.