gpt(eval_full): Read demonstrations from package data
Instead of passing the path to the file, which can be unintuitive, you
can now give the name of the resource in the package instead, with a new
subcommand to print the available files.

This assumes that the files are now frozen as part of the package, but I
was already assuming this for the prompts, so it doesn't seem like a big
deal to me.