util(serde): Add Protocol for Pydantic BaseModel
I want to be able to use protocols for functions that need BaseModels,
but you can't use a class as part of a protocol. This only has
`model_dump` and `model_validate` because it's what I needed to make
`replace_fields` work, but I might add more stuff in the future.