tasks: Use just to manage experiments
The goal is to have a list of different experiment commands that I've
used over time to make them easier to reproduce.

Why just? Mostly because it makes it easier to list everything and add
documentation, which is a little harder in shell scripts. It also
motivates me to keep the commands simple.