exp: Add runs with full data and commands to run all
Now it's possible to run all the experiments with:

    just exp all-base all-full

It's missing commands for 4o instead of 4o-mini because 4o actually
doesn't perform that well and it's too expensive anyway.

Using the full data leads to worse performance. With the base data (n=50
papers), we have ~64% accuracy. With the full data, it's ~56%.