util: Split display_params into more functions
The goal is to both get the params as a dict (to save to a file) and
display them in the program output. By separating into different
functions, we can do that. I still have `display_params`, which now just
calls those functions, so I don't have to fix all callers, although I
will do that at some point.