exp: Manually format root path
I usually prefer automated formatting, but I don't even know if that
exists for Justfiles. So I'm doing this one thing manually, which is hug
the root path variable to the rest of the path.

Ideally, just would have interpolation for top-level strings, but it
doesn't, so I'm doing this instead.