Move common related paper types from `peter` to `related_papers`
This will be used by other paper retrieval implementations.

Note: `related` was a poor name for the module (obvious name for a
variable), but `related_papers` is a little verbose, and it gets used a
lot in some files. So I'm experimenting with `rp` for the shorthand,
like `emb` for `embedding` and `s2` for `semantic_scholar`.