gpt(eval_rationale): Introduce possible alternative output types
As usual: base class with the required methods, a result type that all
output types must convert to, and the final output type `GraphWithEval`
accepts the result type as an argument.

The requirements for output types are: str -> int dict for the actual
metrics, and an explanation text saying why those metrics were given.
This explanation is largely inconsequential, mostly just space for the
model to "think".