gpt(eval_tour): Add more items to PaperCore data
I couldn't include reviews, references, etc. because they deviate in
format between types, so I only added the primitive fields.

main_text wasn't added because it's a property that can take a while to
be generated, and I don't think it's necessary. If it every becomes so,
I'll add it.