gpt: Refactor GPTResult+PromptResult to use a traverse
From functional programming, traverse maps a function over a structure
and sequences the results. The structure here is a list of results
converting to a result of list, and the sequencing is summing the costs
inside the result. Note that for the mapping, we stack GPTResult with
PromptResult to get the inner value mapping.