types: Use mixin to proxy paper properties
This adds properties for the required paper properties in types that
wrap an underlying paper object so I don't have to repeat the same
properties everywhere.

The sad thing is that I can't use `Record` in the proxy because pydantic
doesn't like it, so I have to add the ID manually.