types: Introduce Identifiable protocol and Immutable ABC
Immutable is just BaseModel with frozen=True. Identifiable is the `id`
property as a protocol.

These are the decomposition of Record. This is because I can't have
PaperProxy inherit from Record (pydantic won't let me), but with this
Immutable + PaperProxy setup, I can have PaperProxy define `id` and be
compatible with `Identifiable`.

Now that Immutable only does the BaseModel + frozen part, I'll slowly
replace that combo with it.