Remove random file that got commited `q`
The one problem with jujutsu is situations like these where a random
file get added, and since there's no staging area, it's included with
the rest of the changes.

The commit that introduced this is so far into the past (a week ago),
that I'm just not going to bother.