Replace stdlib's json with orjson
orjson about 3-5 times faster for deserialisation than stdlib's. The
load time for the evaluation scripts is kind of annoying, and this is
part of it.

PS: I didn't benchmark deserialisation because it's not that
important, but it should be faster, too.

PPS: I also did a quick check on other libraries (pysimdjson, ujson,
etc.) and orjson was the fastest. This aligns with [1].

[1] https://github.com/tktech/json_benchmark