s2: PaperFromPeerRead defaults abstract to empty string
There's a paper named 'Graph Structure of Neural Networks' that somehow
only has two extracted citations, and both of them have None abstracts
in S2. This change replaces a None abstract with an empty string.

The alternative is changing the type to `abstract: str|None`, but
everything assumes the abstract is always available.