deps: Use TSeq type instead of Sequence in some cases
`Sequence` is immutable, but accepts types like `list`, so it's not
hashable. However, we _need_ some places here to be hashable, so
`Sequence` is not enough.

This is kind of annoying because the type system does nothing to alert
us that we're using non-hashable types, and our only choice is to see
what panics at runtime. Using the concrete TSeq makes this a little
better.