frontend: Allow "ambiguous" single-line lambdas
I.e. allow lambdas that return void to be used in the single-line
version, instead of forcing brackets always.

While this is ambiguous, it's not enough to offset how _pretty_ the
single-line versions look.

I could've tried to replace `() => func()`, but I think those behave
weirdly in JS and I don't know enough, so they stay.