WIP: Introduce partial evaluation endpoint
Allows the user to give the paper title and abstract, and we retrieve
the recommended papers and find the semantic related papers.

It's partial because without the main paper full contents, we can't
generate citation related papers nor the hiearchical graph.

Note: this is just the scaffolding. The actual evaluation implementation
is WIP.