gpt: Use monadic sequencing of evaluation steps
It looks kind of goofy in Python, but it reduces the change that I leave
a result (especially the cost), behind.

Adding `lift2` was great combining multiple values seems to be a common
occurrence. I'll add more versions if they're required in the future.

Also, I probably need a better name for this. GPTResult isn't worthy of
it.