exp(simple): Create new file with different input data vars
The previous version had different task files for PeerRead and
SciNova/ORC, which is confusing and unnecessary. I think it's better to
do this in a simpler way that's easy to use, even if it requires more
variables to be defined.

The reason for this is that I came back to this after a few months
without touching it and couldn't figure out how to run the PeerRead
experiments.